+++ src/ca/goldorion/mcrpcreator/MainApp.java
@@ -20,7 +20,7 @@ import java.io.IOException;
 
 public class MainApp extends Application {
     public static final String appName = "MCR Plugin Creator";
-    public static final String version = "0.5.0";
+    public static final String version = "0.5.1";
 
     private Stage primaryStage;
     private BorderPane rootLayout;